refactor(core): split lifecycle-manager.ts into focused modules
lifecycle-manager.ts was 2077 lines and mixed polling, state machine,
reaction engine, PR enrichment caching, and backlog dispatchers in a
single file. Split into four focused modules behind the existing
createLifecycleManager factory — no behavior change.

- lifecycle-events.ts: pure helpers (parseDuration, inferPriority,
  createEvent, statusToEventType, prStateToEventType, eventToReactionKey,
  transitionLogLevel, primaryLifecycleReason, buildTransitionObservabilityData,
  makeFingerprint)
- pr-enrichment-cache.ts: per-poll-cycle batch GraphQL cache with
  populate/get/clear
- reaction-engine.ts: executeReaction, retry trackers, escalation logic,
  and reaction config resolution
- lifecycle-backlog.ts: review backlog, CI failure details, and merge
  conflict dispatchers plus throttle

lifecycle-manager.ts shrinks 2077 -> 1191 lines and owns only the polling
loop, state machine transitions (determineStatus, checkSession), merge
auto-cleanup, and report-watcher audit — wiring the extracted modules
together. Public API is unchanged.

Closes #1417
