+++ packages/core/src/lifecycle-manager.ts
@@ -1,16 +1,15 @@
 /**
- * Lifecycle Manager — state machine + polling loop + reaction engine.
+ * Lifecycle Manager — polling loop + state machine + event dispatch.
  *
  * Periodically polls all sessions and:
  * 1. Detects state transitions (spawning → working → pr_open → etc.)
  * 2. Emits events on transitions
- * 3. Triggers reactions (auto-handle CI failures, review comments, etc.)
- * 4. Escalates to human notification when auto-handling fails
+ * 3. Delegates reactions, PR enrichment caching, and backlog dispatch to
+ *    focused modules (reaction-engine, pr-enrichment-cache, lifecycle-backlog).
  *
  * Reference: scripts/claude-session-status, scripts/claude-review-check
  */
 
-import { randomUUID } from "node:crypto";
 import {
   SESSION_STATUS,
   ACTIVITY_STATE,
@@ -21,7 +20,6 @@ import {
   type OpenCodeSessionManager,
   type SessionId,
   type SessionStatus,
-  type EventType,
   type OrchestratorEvent,
   type OrchestratorConfig,
   type ReactionConfig,
@@ -32,14 +30,9 @@ import {
   type SCM,
   type Notifier,
   type Session,
-  type CanonicalSessionLifecycle,
   type EventPriority,
   type ProjectConfig as _ProjectConfig,
-  type PREnrichmentData,
-  type CICheck,
 } from "./types.js";
-import { formatAutomatedCommentsMessage } from "./format-automated-comments.js";
-import { DEFAULT_BUGBOT_COMMENTS_MESSAGE } from "./config.js";
 import { buildLifecycleMetadataPatch, cloneLifecycle, deriveLegacyStatus } from "./lifecycle-state.js";
 import { updateMetadata } from "./metadata.js";
 import { getSessionsDir } from "./paths.js";
@@ -74,156 +67,20 @@ import {
   resolveProbeDecision,
   type LifecycleDecision,
 } from "./lifecycle-status-decisions.js";
-
-/** Parse a duration string like "10m", "30s", "1h" to milliseconds. */
-function parseDuration(str: string): number {
-  const match = str.match(/^(\d+)(s|m|h)$/);
-  if (!match) return 0;
-  const value = parseInt(match[1], 10);
-  switch (match[2]) {
-    case "s":
-      return value * 1000;
-    case "m":
-      return value * 60_000;
-    case "h":
-      return value * 3_600_000;
-    default:
-      return 0;
-  }
-}
-
-/** Infer a reasonable priority from event type. */
-function inferPriority(type: EventType): EventPriority {
-  if (type.includes("stuck") || type.includes("needs_input") || type.includes("errored")) {
-    return "urgent";
-  }
-  if (type.startsWith("summary.")) {
-    return "info";
-  }
-  if (
-    type.includes("approved") ||
-    type.includes("ready") ||
-    type.includes("merged") ||
-    type.includes("completed")
-  ) {
-    return "action";
-  }
-  if (type.includes("fail") || type.includes("changes_requested") || type.includes("conflicts")) {
-    return "warning";
-  }
-  return "info";
-}
-
-/** Create an OrchestratorEvent with defaults filled in. */
-function createEvent(
-  type: EventType,
-  opts: {
-    sessionId: SessionId;
-    projectId: string;
-    message: string;
-    priority?: EventPriority;
-    data?: Record<string, unknown>;
-  },
-): OrchestratorEvent {
-  return {
-    id: randomUUID(),
-    type,
-    priority: opts.priority ?? inferPriority(type),
-    sessionId: opts.sessionId,
-    projectId: opts.projectId,
-    timestamp: new Date(),
-    message: opts.message,
-    data: opts.data ?? {},
-  };
-}
-
-/** Determine which event type corresponds to a status transition. */
-function statusToEventType(_from: SessionStatus | undefined, to: SessionStatus): EventType | null {
-  switch (to) {
-    case "working":
-      return "session.working";
-    case "pr_open":
-      return "pr.created";
-    case "ci_failed":
-      return "ci.failing";
-    case "review_pending":
-      return "review.pending";
-    case "changes_requested":
-      return "review.changes_requested";
-    case "approved":
-      return "review.approved";
-    case "mergeable":
-      return "merge.ready";
-    case "merged":
-      return "merge.completed";
-    case "needs_input":
-      return "session.needs_input";
-    case "stuck":
-      return "session.stuck";
-    case "errored":
-      return "session.errored";
-    case "killed":
-      return "session.killed";
-    default:
-      return null;
-  }
-}
-
-function prStateToEventType(
-  from: Session["lifecycle"]["pr"]["state"],
-  to: Session["lifecycle"]["pr"]["state"],
-): EventType | null {
-  if (from === to) return null;
-  switch (to) {
-    case "closed":
-      return "pr.closed";
-    default:
-      return null;
-  }
-}
-
-/** Map event type to reaction config key. */
-function eventToReactionKey(eventType: EventType): string | null {
-  switch (eventType) {
-    case "pr.closed":
-      return "pr-closed";
-    case "ci.failing":
-      return "ci-failed";
-    case "review.changes_requested":
-      return "changes-requested";
-    case "automated_review.found":
-      return "bugbot-comments";
-    case "merge.conflicts":
-      return "merge-conflicts";
-    case "merge.ready":
-      return "approved-and-green";
-    case "session.stuck":
-      return "agent-stuck";
-    case "session.needs_input":
-      return "agent-needs-input";
-    case "session.killed":
-      return "agent-exited";
-    case "summary.all_complete":
-      return "all-complete";
-    default:
-      return null;
-  }
-}
-
-function transitionLogLevel(status: SessionStatus): "info" | "warn" | "error" {
-  const eventType = statusToEventType(undefined, status);
-  if (!eventType) {
-    return "info";
-  }
-  const priority = inferPriority(eventType);
-  if (priority === "urgent") {
-    return "error";
-  }
-  if (priority === "warning") {
-    return "warn";
-  }
-  return "info";
-}
+import {
+  buildTransitionObservabilityData,
+  createEvent,
+  eventToReactionKey,
+  inferPriority,
+  parseDuration,
+  prStateToEventType,
+  primaryLifecycleReason,
+  statusToEventType,
+  transitionLogLevel,
+} from "./lifecycle-events.js";
+import { createPREnrichmentCache } from "./pr-enrichment-cache.js";
+import { createReactionEngine } from "./reaction-engine.js";
+import { createBacklogDispatchers } from "./lifecycle-backlog.js";
 
 interface DeterminedStatus {
   status: SessionStatus;
@@ -240,61 +97,6 @@ interface ProbeResult {
   failed: boolean;
 }
 
-function splitEvidenceSignals(evidence: string): string[] {
-  return evidence
-    .split(/\s+/)
-    .map((signal) => signal.trim())
-    .filter((signal) => signal.length > 0);
-}
-
-function primaryLifecycleReason(lifecycle: CanonicalSessionLifecycle): string {
-  if (lifecycle.session.state === "detecting") return lifecycle.session.reason;
-  if (lifecycle.pr.reason !== "not_created" && lifecycle.pr.reason !== "in_progress") {
-    return lifecycle.pr.reason;
-  }
-  if (lifecycle.runtime.reason !== "process_running") {
-    return lifecycle.runtime.reason;
-  }
-  return lifecycle.session.reason;
-}
-
-function buildTransitionObservabilityData(
-  previous: CanonicalSessionLifecycle,
-  next: CanonicalSessionLifecycle,
-  oldStatus: SessionStatus,
-  newStatus: SessionStatus,
-  evidence: string,
-  detectingAttempts: number,
-  statusTransition: boolean,
-  reaction?: { key: string; result: ReactionResult | null },
-): Record<string, unknown> {
-  return {
-    oldStatus,
-    newStatus,
-    statusTransition,
-    previousSessionState: previous.session.state,
-    newSessionState: next.session.state,
-    previousSessionReason: previous.session.reason,
-    newSessionReason: next.session.reason,
-    previousPRState: previous.pr.state,
-    newPRState: next.pr.state,
-    previousPRReason: previous.pr.reason,
-    newPRReason: next.pr.reason,
-    previousRuntimeState: previous.runtime.state,
-    newRuntimeState: next.runtime.state,
-    previousRuntimeReason: previous.runtime.reason,
-    newRuntimeReason: next.runtime.reason,
-    primaryReason: primaryLifecycleReason(next),
-    evidence,
-    signalsConsulted: splitEvidenceSignals(evidence),
-    detectingAttempts,
-    recoveryAction: reaction?.result?.action ?? null,
-    reactionKey: reaction?.key ?? null,
-    reactionSuccess: reaction?.result?.success ?? null,
-    escalated: reaction?.result?.escalated ?? null,
-  };
-}
-
 export interface LifecycleManagerDeps {
   config: OrchestratorConfig;
   registry: PluginRegistry;
@@ -303,153 +105,42 @@ export interface LifecycleManagerDeps {
   projectId?: string;
 }
 
-/** Track attempt counts for reactions per session. */
-interface ReactionTracker {
-  attempts: number;
-  firstTriggered: Date;
-}
-
 /** Create a LifecycleManager instance. */
 export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleManager {
   const { config, registry, sessionManager, projectId: scopedProjectId } = deps;
   const observer = createProjectObserver(config, "lifecycle-manager");
 
   const states = new Map<SessionId, SessionStatus>();
-  const reactionTrackers = new Map<string, ReactionTracker>(); // "sessionId:reactionKey"
   let pollTimer: ReturnType<typeof setInterval> | null = null;
   let polling = false; // re-entrancy guard
   let allCompleteEmitted = false; // guard against repeated all_complete
 
-  /**
-   * Cache for PR enrichment data within a single poll cycle.
-   * Cleared at the start of each pollAll() call.
-   * Key format: "${owner}/${repo}#${number}"
-   */
-  const prEnrichmentCache = new Map<string, PREnrichmentData>();
+  const prEnrichmentCache = createPREnrichmentCache({
+    config,
+    registry,
+    observer,
+    scopedProjectId,
+  });
 
-  /**
-   * Per-session timestamp of last review backlog API check.
-   * Used to throttle getPendingComments/getAutomatedComments to at most once per 2 minutes.
-   * In-memory only — resets on restart (acceptable since it's a rate-limit hint, not state).
-   */
-  const lastReviewBacklogCheckAt = new Map<SessionId, number>();
+  const reactionEngine = createReactionEngine({
+    config,
+    sessionManager,
+    notifyHuman,
+  });
 
-  /** Throttle interval for review backlog API calls (2 minutes). */
-  const REVIEW_BACKLOG_THROTTLE_MS = 2 * 60 * 1000;
+  const backlog = createBacklogDispatchers({
+    config,
+    registry,
+    sessionManager,
+    reactionEngine,
+    prEnrichmentCache,
+    updateSessionMetadata,
+    notifyHuman,
+  });
 
-  /**
-   * Populate the PR enrichment cache using batch GraphQL queries.
-   * This is called once per poll cycle to fetch data for all PRs efficiently.
-   */
-  async function populatePREnrichmentCache(sessions: Session[]): Promise<void> {
-    // Clear previous cache
-    prEnrichmentCache.clear();
-
-    // Collect all unique PRs keyed by their owning session's project/plugin.
-    const prsByPlugin = new Map<string, Array<NonNullable<Session["pr"]>>>();
-    const seenPRKeys = new Set<string>();
-    for (const session of sessions) {
-      if (!session.pr) continue;
-      const project = config.projects[session.projectId];
-      if (!project?.scm?.plugin) continue;
-
-      const prKey = `${session.pr.owner}/${session.pr.repo}#${session.pr.number}`;
-      if (seenPRKeys.has(prKey)) continue;
-      seenPRKeys.add(prKey);
-
-      const pluginKey = project.scm.plugin;
-      if (!prsByPlugin.has(pluginKey)) {
-        prsByPlugin.set(pluginKey, []);
-      }
-      const pluginPRs = prsByPlugin.get(pluginKey);
-      if (pluginPRs) {
-        pluginPRs.push(session.pr);
-      }
-    }
-
-    // Fetch enrichment data for each plugin's PRs
-    for (const [pluginKey, pluginPRs] of prsByPlugin) {
-      const scm = registry.get<SCM>("scm", pluginKey);
-      if (!scm?.enrichSessionsPRBatch) continue;
-
-      const batchStartTime = Date.now();
-      try {
-        const enrichmentData = await scm.enrichSessionsPRBatch(
-          pluginPRs,
-          {
-            recordSuccess(_data) {
-              const batchDuration = Date.now() - batchStartTime;
-              observer?.recordOperation({
-                metric: "graphql_batch",
-                operation: "batch_enrichment",
-                correlationId: createCorrelationId("graphql-batch"),
-                outcome: "success",
-                projectId: scopedProjectId,
-                durationMs: batchDuration,
-                data: {
-                  plugin: pluginKey,
-                  prCount: pluginPRs.length,
-                  prKeys: pluginPRs.map((pr) => `${pr.owner}/${pr.repo}#${pr.number}`),
-                },
-                level: "info",
-              });
-            },
-            recordFailure(data) {
-              const batchDuration = Date.now() - batchStartTime;
-              observer?.recordOperation({
-                metric: "graphql_batch",
-                operation: "batch_enrichment",
-                correlationId: createCorrelationId("graphql-batch"),
-                outcome: "failure",
-                reason: data.error,
-                level: "warn",
-                data: {
-                  plugin: pluginKey,
-                  prCount: pluginPRs.length,
-                  error: data.error,
-                  durationMs: batchDuration,
-                },
-              });
-            },
-            log(level, message) {
-              observer?.recordDiagnostic?.({
-                operation: "batch_enrichment.log",
-                correlationId: createCorrelationId("graphql-batch"),
-                projectId: scopedProjectId,
-                message,
-                level,
-                data: {
-                  plugin: pluginKey,
-                  source: "ao-graphql-batch",
-                },
-              });
-            },
-          },
-        );
-
-        // Merge into cache
-        for (const [key, data] of enrichmentData) {
-          prEnrichmentCache.set(key, data);
-        }
-      } catch (err) {
-        // Batch fetch failed - individual calls will still work
-        const errorMsg = err instanceof Error ? err.message : String(err);
-        const batchCorrelationId = createCorrelationId("batch-enrichment");
-        observer?.recordOperation?.({
-          metric: "lifecycle_poll",
-          operation: "batch_enrichment",
-          correlationId: batchCorrelationId,
-          outcome: "failure",
-          reason: errorMsg,
-          level: "warn",
-          data: { plugin: pluginKey, prCount: pluginPRs.length },
-        });
-      }
-    }
-  }
   /** Check if idle time exceeds the agent-stuck threshold. */
   function isIdleBeyondThreshold(session: Session, idleTimestamp: Date): boolean {
-    const stuckReaction = getReactionConfigForSession(session, "agent-stuck");
+    const stuckReaction = reactionEngine.getReactionConfigForSession(session, "agent-stuck");
     const thresholdStr = stuckReaction?.threshold;
     if (typeof thresholdStr !== "string") return false;
     const stuckThresholdMs = parseDuration(thresholdStr);
@@ -910,150 +601,6 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
     });
   }
 
-  /** Execute a reaction for a session. */
-  async function executeReaction(
-    sessionId: SessionId,
-    projectId: string,
-    reactionKey: string,
-    reactionConfig: ReactionConfig,
-  ): Promise<ReactionResult> {
-    const trackerKey = `${sessionId}:${reactionKey}`;
-    let tracker = reactionTrackers.get(trackerKey);
-
-    if (!tracker) {
-      tracker = { attempts: 0, firstTriggered: new Date() };
-      reactionTrackers.set(trackerKey, tracker);
-    }
-
-    // Increment attempts before checking escalation
-    tracker.attempts++;
-
-    // Check if we should escalate
-    const maxRetries = reactionConfig.retries ?? Infinity;
-    const escalateAfter = reactionConfig.escalateAfter;
-    let shouldEscalate = false;
-
-    if (tracker.attempts > maxRetries) {
-      shouldEscalate = true;
-    }
-
-    if (typeof escalateAfter === "string") {
-      const durationMs = parseDuration(escalateAfter);
-      if (durationMs > 0 && Date.now() - tracker.firstTriggered.getTime() > durationMs) {
-        shouldEscalate = true;
-      }
-    }
-
-    if (typeof escalateAfter === "number" && tracker.attempts > escalateAfter) {
-      shouldEscalate = true;
-    }
-
-    if (shouldEscalate) {
-      // Escalate to human
-      const event = createEvent("reaction.escalated", {
-        sessionId,
-        projectId,
-        message: `Reaction '${reactionKey}' escalated after ${tracker.attempts} attempts`,
-        data: { reactionKey, attempts: tracker.attempts },
-      });
-      await notifyHuman(event, reactionConfig.priority ?? "urgent");
-      return {
-        reactionType: reactionKey,
-        success: true,
-        action: "escalated",
-        escalated: true,
-      };
-    }
-
-    // Execute the reaction action
-    const action = reactionConfig.action ?? "notify";
-
-    switch (action) {
-      case "send-to-agent": {
-        if (reactionConfig.message) {
-          try {
-            await sessionManager.send(sessionId, reactionConfig.message);
-
-            return {
-              reactionType: reactionKey,
-              success: true,
-              action: "send-to-agent",
-              message: reactionConfig.message,
-              escalated: false,
-            };
-          } catch {
-            // Send failed — allow retry on next poll cycle (don't escalate immediately)
-            return {
-              reactionType: reactionKey,
-              success: false,
-              action: "send-to-agent",
-              escalated: false,
-            };
-          }
-        }
-        break;
-      }
-
-      case "notify": {
-        const event = createEvent("reaction.triggered", {
-          sessionId,
-          projectId,
-          message: `Reaction '${reactionKey}' triggered notification`,
-          data: { reactionKey },
-        });
-        await notifyHuman(event, reactionConfig.priority ?? "info");
-        return {
-          reactionType: reactionKey,
-          success: true,
-          action: "notify",
-          escalated: false,
-        };
-      }
-
-      case "auto-merge": {
-        // Auto-merge is handled by the SCM plugin
-        // For now, just notify
-        const event = createEvent("reaction.triggered", {
-          sessionId,
-          projectId,
-          message: `Reaction '${reactionKey}' triggered auto-merge`,
-          data: { reactionKey },
-        });
-        await notifyHuman(event, "action");
-        return {
-          reactionType: reactionKey,
-          success: true,
-          action: "auto-merge",
-          escalated: false,
-        };
-      }
-    }
-
-    return {
-      reactionType: reactionKey,
-      success: false,
-      action,
-      escalated: false,
-    };
-  }
-
-  function clearReactionTracker(sessionId: SessionId, reactionKey: string): void {
-    reactionTrackers.delete(`${sessionId}:${reactionKey}`);
-  }
-
-  function getReactionConfigForSession(
-    session: Session,
-    reactionKey: string,
-  ): ReactionConfig | null {
-    const project = config.projects[session.projectId];
-    const globalReaction = config.reactions[reactionKey];
-    const projectReaction = project?.reactions?.[reactionKey];
-    const reactionConfig = projectReaction
-      ? { ...globalReaction, ...projectReaction }
-      : globalReaction;
-    return reactionConfig ? (reactionConfig as ReactionConfig) : null;
-  }
-
   function updateSessionMetadata(
     session: Session,
     updates: Partial<Record<string, string>>,
@@ -1082,457 +629,6 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
     }
     session.metadata = cleaned;
     session.status = deriveLegacyStatus(session.lifecycle, session.status);
-  }
-
-  function makeFingerprint(ids: string[]): string {
-    return [...ids].sort().join(",");
-  }
-
-  async function maybeDispatchReviewBacklog(
-    session: Session,
-    oldStatus: SessionStatus,
-    newStatus: SessionStatus,
-    transitionReaction?: { key: string; result: ReactionResult | null },
-  ): Promise<void> {
-    const project = config.projects[session.projectId];
-    if (!project || !session.pr) return;
-
-    const scm = project.scm?.plugin ? registry.get<SCM>("scm", project.scm.plugin) : null;
-    if (!scm) return;
-
-    const humanReactionKey = "changes-requested";
-    const automatedReactionKey = "bugbot-comments";
-
-    if (TERMINAL_STATUSES.has(newStatus) || session.lifecycle.pr.state !== "open") {
-      clearReactionTracker(session.id, humanReactionKey);
-      clearReactionTracker(session.id, automatedReactionKey);
-      lastReviewBacklogCheckAt.delete(session.id);
-      updateSessionMetadata(session, {
-        lastPendingReviewFingerprint: "",
-        lastPendingReviewDispatchHash: "",
-        lastPendingReviewDispatchAt: "",
-        lastAutomatedReviewFingerprint: "",
-        lastAutomatedReviewDispatchHash: "",
-        lastAutomatedReviewDispatchAt: "",
-      });
-      return;
-    }
-
-    // Throttle review backlog API calls to at most once per 2 minutes.
-    // Comments don't change faster than this in practice, and the SCM calls
-    // (getPendingComments + getAutomatedComments) consume API quota on every poll.
-    //
-    // Exception: bypass throttle when a transition reaction just fired for a
-    // review reaction key. The transitionReaction branch records
-    // lastPendingReviewDispatchHash, which requires the current fingerprint from
-    // the API. If we throttle here, that metadata never gets written and the
-    // next unthrottled poll sees a "new" fingerprint, clears the reaction tracker,
-    // and fires a duplicate dispatch.
-    const hasRelevantTransition =
-      transitionReaction?.key === humanReactionKey ||
-      transitionReaction?.key === automatedReactionKey;
-    if (!hasRelevantTransition) {
-      const lastCheckAt = lastReviewBacklogCheckAt.get(session.id) ?? 0;
-      if (Date.now() - lastCheckAt < REVIEW_BACKLOG_THROTTLE_MS) {
-        return;
-      }
-    }
-    lastReviewBacklogCheckAt.set(session.id, Date.now());
-
-    const [pendingResult, automatedResult] = await Promise.allSettled([
-      scm.getPendingComments(session.pr),
-      scm.getAutomatedComments(session.pr),
-    ]);
-
-    // null means "failed to fetch" — preserve existing metadata.
-    // [] means "confirmed no comments" — safe to clear.
-    const pendingComments =
-      pendingResult.status === "fulfilled" && Array.isArray(pendingResult.value)
-        ? pendingResult.value
-        : null;
-    const automatedComments =
-      automatedResult.status === "fulfilled" && Array.isArray(automatedResult.value)
-        ? automatedResult.value
-        : null;
-
-    // --- Pending (human) review comments ---
-    // null = SCM fetch failed; skip processing to preserve existing metadata.
-    if (pendingComments !== null) {
-      const pendingFingerprint = makeFingerprint(pendingComments.map((comment) => comment.id));
-      const lastPendingFingerprint = session.metadata["lastPendingReviewFingerprint"] ?? "";
-      const lastPendingDispatchHash = session.metadata["lastPendingReviewDispatchHash"] ?? "";
-
-      if (
-        pendingFingerprint !== lastPendingFingerprint &&
-        transitionReaction?.key !== humanReactionKey
-      ) {
-        clearReactionTracker(session.id, humanReactionKey);
-      }
-      if (pendingFingerprint !== lastPendingFingerprint) {
-        updateSessionMetadata(session, {
-          lastPendingReviewFingerprint: pendingFingerprint,
-        });
-      }
-
-      if (!pendingFingerprint) {
-        clearReactionTracker(session.id, humanReactionKey);
-        updateSessionMetadata(session, {
-          lastPendingReviewFingerprint: "",
-          lastPendingReviewDispatchHash: "",
-          lastPendingReviewDispatchAt: "",
-        });
-      } else if (
-        transitionReaction?.key === humanReactionKey &&
-        transitionReaction.result?.success
-      ) {
-        if (lastPendingDispatchHash !== pendingFingerprint) {
-          updateSessionMetadata(session, {
-            lastPendingReviewDispatchHash: pendingFingerprint,
-            lastPendingReviewDispatchAt: new Date().toISOString(),
-          });
-        }
-      } else if (
-        !(oldStatus !== newStatus && newStatus === "changes_requested") &&
-        pendingFingerprint !== lastPendingDispatchHash
-      ) {
-        const reactionConfig = getReactionConfigForSession(session, humanReactionKey);
-        if (
-          reactionConfig &&
-          reactionConfig.action &&
-          (reactionConfig.auto !== false || reactionConfig.action === "notify")
-        ) {
-          const result = await executeReaction(
-            session.id,
-            session.projectId,
-            humanReactionKey,
-            reactionConfig,
-          );
-          if (result.success) {
-            updateSessionMetadata(session, {
-              lastPendingReviewDispatchHash: pendingFingerprint,
-              lastPendingReviewDispatchAt: new Date().toISOString(),
-            });
-          }
-        }
-      }
-    }
-
-    // --- Automated (bot) review comments ---
-    if (automatedComments !== null) {
-      const automatedFingerprint = makeFingerprint(
-        automatedComments.map((comment) => comment.id),
-      );
-      const lastAutomatedFingerprint = session.metadata["lastAutomatedReviewFingerprint"] ?? "";
-      const lastAutomatedDispatchHash =
-        session.metadata["lastAutomatedReviewDispatchHash"] ?? "";
-
-      if (automatedFingerprint !== lastAutomatedFingerprint) {
-        clearReactionTracker(session.id, automatedReactionKey);
-        updateSessionMetadata(session, {
-          lastAutomatedReviewFingerprint: automatedFingerprint,
-        });
-      }
-
-      if (!automatedFingerprint) {
-        clearReactionTracker(session.id, automatedReactionKey);
-        updateSessionMetadata(session, {
-          lastAutomatedReviewFingerprint: "",
-          lastAutomatedReviewDispatchHash: "",
-          lastAutomatedReviewDispatchAt: "",
-        });
-      } else if (automatedFingerprint !== lastAutomatedDispatchHash) {
-        const reactionConfig = getReactionConfigForSession(session, automatedReactionKey);
-        if (
-          reactionConfig &&
-          reactionConfig.action &&
-          (reactionConfig.auto !== false || reactionConfig.action === "notify")
-        ) {
-          // Inject the detailed comment listing + correct-API guidance into the
-          // message so the agent doesn't re-fetch with stale or unpaginated calls
-          // (see #895 — fixes the pagination + stale `gh pr checks` failure modes).
-          // Only override when the message is the built-in sentinel — a user who
-          // customized `reactions.bugbot-comments.message` in their YAML gets
-          // exactly what they wrote, nothing more.
-          const usingDefaultMessage =
-            reactionConfig.message === DEFAULT_BUGBOT_COMMENTS_MESSAGE;
-          const detailedConfig: ReactionConfig =
-            reactionConfig.action === "send-to-agent" && usingDefaultMessage
-              ? {
-                  ...reactionConfig,
-                  message: formatAutomatedCommentsMessage(automatedComments, session.pr),
-                }
-              : reactionConfig;
-          const result = await executeReaction(
-            session.id,
-            session.projectId,
-            automatedReactionKey,
-            detailedConfig,
-          );
-          if (result.success) {
-            updateSessionMetadata(session, {
-              lastAutomatedReviewDispatchHash: automatedFingerprint,
-              lastAutomatedReviewDispatchAt: new Date().toISOString(),
-            });
-          }
-        }
-      }
-    }
-  }
-
-  /**
-   * Format CI check failures into a human-readable message for the agent.
-   * Includes check names, statuses, and links for debugging.
-   */
-  function formatCIFailureMessage(failedChecks: CICheck[]): string {
-    const lines = [
-      "CI checks are failing on your PR. Here are the failed checks:",
-      "",
-    ];
-    for (const check of failedChecks) {
-      const status = check.conclusion ?? check.status;
-      const link = check.url ? ` — ${check.url}` : "";
-      lines.push(`- **${check.name}**: ${status}${link}`);
-    }
-    lines.push(
-      "",
-      "Investigate the failures, fix the issues, and push again.",
-    );
-    return lines.join("\n");
-  }
-
-  /**
-   * Dispatch CI failure details to the agent session when new or changed
-   * failures are detected. Follows the same fingerprinting/deduplication
-   * pattern as maybeDispatchReviewBacklog().
-   */
-  async function maybeDispatchCIFailureDetails(
-    session: Session,
-    _oldStatus: SessionStatus,
-    newStatus: SessionStatus,
-    transitionReaction?: { key: string; result: ReactionResult | null },
-  ): Promise<void> {
-    const project = config.projects[session.projectId];
-    if (!project || !session.pr) return;
-
-    const scm = project.scm?.plugin ? registry.get<SCM>("scm", project.scm.plugin) : null;
-    if (!scm) return;
-
-    const ciReactionKey = "ci-failed";
-
-    // Clear tracking when PR is closed/merged
-    if (newStatus === "merged" || newStatus === "killed") {
-      clearReactionTracker(session.id, ciReactionKey);
-      updateSessionMetadata(session, {
-        lastCIFailureFingerprint: "",
-        lastCIFailureDispatchHash: "",
-        lastCIFailureDispatchAt: "",
-      });
-      return;
-    }
-
-    // Only dispatch CI details when in ci_failed state
-    if (newStatus !== "ci_failed") {
-      // CI is no longer failing — clear tracking so next failure is dispatched fresh
-      const lastFingerprint = session.metadata["lastCIFailureFingerprint"] ?? "";
-      if (lastFingerprint) {
-        clearReactionTracker(session.id, ciReactionKey);
-        updateSessionMetadata(session, {
-          lastCIFailureFingerprint: "",
-          lastCIFailureDispatchHash: "",
-          lastCIFailureDispatchAt: "",
-        });
-      }
-      return;
-    }
-
-    // Fetch individual CI checks for failure details.
-    // Use batch enrichment data when available to avoid an extra REST call;
-    // fall back to getCIChecks() when the batch didn't run this cycle.
-    const prKey = `${session.pr.owner}/${session.pr.repo}#${session.pr.number}`;
-    const cachedEnrichment = prEnrichmentCache.get(prKey);
-
-    let checks: CICheck[];
-    if (cachedEnrichment?.ciChecks !== undefined) {
-      checks = cachedEnrichment.ciChecks;
-    } else {
-      try {
-        checks = await scm.getCIChecks(session.pr);
-      } catch {
-        // Failed to fetch checks — skip this cycle
-        return;
-      }
-    }
-
-    const failedChecks = checks.filter(
-      (c) => c.status === "failed" || c.conclusion?.toUpperCase() === "FAILURE",
-    );
-    if (failedChecks.length === 0) return;
-
-    const ciFingerprint = makeFingerprint(
-      failedChecks.map((c) => `${c.name}:${c.status}:${c.conclusion ?? ""}`),
-    );
-    const lastCIFingerprint = session.metadata["lastCIFailureFingerprint"] ?? "";
-    const lastCIDispatchHash = session.metadata["lastCIFailureDispatchHash"] ?? "";
-
-    // Reset reaction tracker when failure set changes
-    if (ciFingerprint !== lastCIFingerprint && transitionReaction?.key !== ciReactionKey) {
-      clearReactionTracker(session.id, ciReactionKey);
-    }
-    if (ciFingerprint !== lastCIFingerprint) {
-      updateSessionMetadata(session, {
-        lastCIFailureFingerprint: ciFingerprint,
-      });
-    }
-
-    // If transition already sent a ci-failed reaction with the static message,
-    // skip this cycle but do NOT record dispatch hash — the next poll will send
-    // the detailed CI failure info with check names and URLs.
-    if (
-      transitionReaction?.key === ciReactionKey &&
-      transitionReaction.result?.success
-    ) {
-      return;
-    }
-
-    // Skip if we already dispatched this exact failure set
-    if (ciFingerprint === lastCIDispatchHash) return;
-
-    // Dispatch CI failure details directly via sessionManager.send() rather than
-    // executeReaction() to avoid consuming the ci-failed reaction's retry budget.
-    // The transition reaction owns escalation; this is a follow-up info delivery.
-    const reactionConfig = getReactionConfigForSession(session, ciReactionKey);
-    if (
-      reactionConfig &&
-      reactionConfig.action &&
-      (reactionConfig.auto !== false || reactionConfig.action === "notify")
-    ) {
-      const detailedMessage = formatCIFailureMessage(failedChecks);
-
-      try {
-        if (reactionConfig.action === "send-to-agent") {
-          await sessionManager.send(session.id, detailedMessage);
-        } else {
-          // For "notify" action, send to human notifiers instead
-          const event = createEvent("ci.failing", {
-            sessionId: session.id,
-            projectId: session.projectId,
-            message: detailedMessage,
-            data: { failedChecks: failedChecks.map((c) => c.name) },
-          });
-          await notifyHuman(event, reactionConfig.priority ?? "warning");
-        }
-
-        updateSessionMetadata(session, {
-          lastCIFailureDispatchHash: ciFingerprint,
-          lastCIFailureDispatchAt: new Date().toISOString(),
-        });
-      } catch {
-        // Send failed — will retry on next poll cycle
-      }
-    }
-  }
-
-  /**
-   * Dispatch merge conflict notifications to the agent session.
-   * Conflicts are detected from the PR enrichment cache or getMergeability()
-   * and dispatched independently of the session status (conflicts can coexist
-   * with ci_failed, changes_requested, etc.).
-   */
-  async function maybeDispatchMergeConflicts(
-    session: Session,
-    newStatus: SessionStatus,
-  ): Promise<void> {
-    const project = config.projects[session.projectId];
-    if (!project || !session.pr) return;
-
-    const scm = project.scm?.plugin ? registry.get<SCM>("scm", project.scm.plugin) : null;
-    if (!scm) return;
-
-    const conflictReactionKey = "merge-conflicts";
-
-    // Clear tracking when PR is no longer open.
-    if (session.lifecycle.pr.state !== "open" || newStatus === "killed") {
-      clearReactionTracker(session.id, conflictReactionKey);
-      updateSessionMetadata(session, {
-        lastMergeConflictDispatched: "",
-      });
-      return;
-    }
-
-    // Only check for conflicts on open PRs
-    if (
-      newStatus !== "pr_open" &&
-      newStatus !== "ci_failed" &&
-      newStatus !== "review_pending" &&
-      newStatus !== "changes_requested" &&
-      newStatus !== "approved" &&
-      newStatus !== "mergeable"
-    ) {
-      return;
-    }
-
-    // Check for conflicts using cached enrichment data or fallback to individual call.
-    // When batch enrichment ran (cachedData is present), use its hasConflicts value
-    // to avoid 3 redundant REST calls from getMergeability() — the batch already
-    // fetched the mergeable/mergeStateStatus fields via GraphQL.
-    const prKey = `${session.pr.owner}/${session.pr.repo}#${session.pr.number}`;
-    const cachedData = prEnrichmentCache.get(prKey);
-
-    let hasConflicts: boolean;
-    if (cachedData) {
-      // Batch ran — trust its data (undefined means CONFLICTING wasn't set → no conflicts)
-      hasConflicts = cachedData.hasConflicts ?? false;
-    } else {
-      // Batch didn't run this cycle — fall back to individual API call
-      try {
-        const mergeReadiness = await scm.getMergeability(session.pr);
-        hasConflicts = !mergeReadiness.noConflicts;
-      } catch {
-        return;
-      }
-    }
-
-    const lastDispatched = session.metadata["lastMergeConflictDispatched"] ?? "";
-
-    if (hasConflicts) {
-      // Already dispatched for current conflict state — skip
-      if (lastDispatched === "true") return;
-
-      const reactionConfig = getReactionConfigForSession(session, conflictReactionKey);
-      if (
-        reactionConfig &&
-        reactionConfig.action &&
-        (reactionConfig.auto !== false || reactionConfig.action === "notify")
-      ) {
-        try {
-          if (reactionConfig.action === "send-to-agent") {
-            const message =
-              reactionConfig.message ??
-              "Your branch has merge conflicts. Rebase on the default branch and resolve them.";
-            await sessionManager.send(session.id, message);
-          } else {
-            const event = createEvent("merge.conflicts", {
-              sessionId: session.id,
-              projectId: session.projectId,
-              message: `${session.id}: PR has merge conflicts`,
-            });
-            await notifyHuman(event, reactionConfig.priority ?? "warning");
-          }
-
-          updateSessionMetadata(session, {
-            lastMergeConflictDispatched: "true",
-          });
-        } catch {
-          // Send failed — will retry on next poll cycle
-        }
-      }
-    } else if (lastDispatched === "true") {
-      // Conflicts resolved — clear so we can re-dispatch if they recur
-      clearReactionTracker(session.id, conflictReactionKey);
-      updateSessionMetadata(session, {
-        lastMergeConflictDispatched: "",
-      });
-    }
   }
 
   /** Send a notification to all configured notifiers. */
@@ -1731,7 +827,7 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
       if (oldEventType) {
         const oldReactionKey = eventToReactionKey(oldEventType);
         if (oldReactionKey) {
-          clearReactionTracker(session.id, oldReactionKey);
+          reactionEngine.clearTracker(session.id, oldReactionKey);
         }
       }
 
@@ -1742,12 +838,12 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
         const reactionKey = eventToReactionKey(eventType);
 
         if (reactionKey) {
-          const reactionConfig = getReactionConfigForSession(session, reactionKey);
+          const reactionConfig = reactionEngine.getReactionConfigForSession(session, reactionKey);
 
           if (reactionConfig && reactionConfig.action) {
             // auto: false skips automated agent actions but still allows notifications
             if (reactionConfig.auto !== false || reactionConfig.action === "notify") {
-              const reactionResult = await executeReaction(
+              const reactionResult = await reactionEngine.executeReaction(
                 session.id,
                 session.projectId,
                 reactionKey,
@@ -1830,10 +926,15 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
       const reactionKey = eventToReactionKey(prEventType);
 
       if (reactionKey) {
-        const reactionConfig = getReactionConfigForSession(session, reactionKey);
+        const reactionConfig = reactionEngine.getReactionConfigForSession(session, reactionKey);
         if (reactionConfig && reactionConfig.action) {
           if (reactionConfig.auto !== false || reactionConfig.action === "notify") {
-            await executeReaction(session.id, session.projectId, reactionKey, reactionConfig);
+            await reactionEngine.executeReaction(
+              session.id,
+              session.projectId,
+              reactionKey,
+              reactionConfig,
+            );
             reactionHandledNotify = true;
           }
         }
@@ -1872,9 +973,9 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
     }
 
     await Promise.allSettled([
-      maybeDispatchReviewBacklog(session, oldStatus, newStatus, transitionReaction),
-      maybeDispatchCIFailureDetails(session, oldStatus, newStatus, transitionReaction),
-      maybeDispatchMergeConflicts(session, newStatus),
+      backlog.maybeDispatchReviewBacklog(session, oldStatus, newStatus, transitionReaction),
+      backlog.maybeDispatchCIFailureDetails(session, oldStatus, newStatus, transitionReaction),
+      backlog.maybeDispatchMergeConflicts(session, newStatus),
     ]);
 
     // Report watcher: audit agent reports for issues (#140)
@@ -1909,7 +1010,7 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
     }
 
     const reactionKey = getReactionKeyForTrigger(auditResult.trigger);
-    const reactionConfig = getReactionConfigForSession(session, reactionKey);
+    const reactionConfig = reactionEngine.getReactionConfigForSession(session, reactionKey);
 
     // Update audit metadata
     const currentTriggerCount = parseInt(
@@ -1951,7 +1052,7 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
 
     // Execute reaction if configured
     if (isNewTrigger && reactionConfig && reactionConfig.auto !== false) {
-      await executeReaction(session.id, session.projectId, reactionKey, reactionConfig);
+      await reactionEngine.executeReaction(session.id, session.projectId, reactionKey, reactionConfig);
     }
   }
 
@@ -1977,12 +1078,12 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
 
       // Prime the per-poll PR enrichment cache before session checks so
       // downstream status/reaction logic can reuse batch GraphQL data.
-      await populatePREnrichmentCache(sessionsToCheck);
+      await prEnrichmentCache.populate(sessionsToCheck);
 
       // Poll all sessions concurrently
       await Promise.allSettled(sessionsToCheck.map((s) => checkSession(s)));
 
-      // Prune stale entries from states, reactionTrackers, and lastReviewBacklogCheckAt
+      // Prune stale entries from states, reactionTrackers, and the review-backlog throttle
       // for sessions that no longer appear in the session list (e.g., after kill/cleanup)
       const currentSessionIds = new Set(sessions.map((s) => s.id));
       for (const trackedId of states.keys()) {
@@ -1990,17 +1091,8 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
           states.delete(trackedId);
         }
       }
-      for (const trackerKey of reactionTrackers.keys()) {
-        const sessionId = trackerKey.split(":")[0];
-        if (sessionId && !currentSessionIds.has(sessionId)) {
-          reactionTrackers.delete(trackerKey);
-        }
-      }
-      for (const sessionId of lastReviewBacklogCheckAt.keys()) {
-        if (!currentSessionIds.has(sessionId)) {
-          lastReviewBacklogCheckAt.delete(sessionId);
-        }
-      }
+      reactionEngine.pruneTrackers(currentSessionIds);
+      backlog.pruneReviewBacklogThrottle(currentSessionIds);
 
       // Check if all sessions are complete (trigger reaction only once)
       const activeSessions = sessions.filter((s) => !TERMINAL_STATUSES.has(s.status));
@@ -2013,7 +1105,12 @@ export function createLifecycleManager(deps: LifecycleManagerDeps): LifecycleMan
           const reactionConfig = config.reactions[reactionKey];
           if (reactionConfig && reactionConfig.action) {
             if (reactionConfig.auto !== false || reactionConfig.action === "notify") {
-              await executeReaction("system", "all", reactionKey, reactionConfig as ReactionConfig);
+              await reactionEngine.executeReaction(
+                "system",
+                "all",
+                reactionKey,
+                reactionConfig as ReactionConfig,
+              );
             }
           }
         }
